feature:1.0.6 增加余额查询接口，支持gpt-4模型，序列化

+++ src/main/java/com/unfbx/chatgpt/entity/chat/ChatCompletion.java
@@ -44,7 +44,7 @@ public class ChatCompletion implements Serializable {
      */
     @JsonProperty("top_p")
     @Builder.Default
-    private Integer topP = 1;
+    private Double topP = 1d;
 
 
     /**
@@ -129,7 +129,7 @@ public class ChatCompletion implements Serializable {
         this.stop = stop;
     }
 
-    public void setTopP(Integer topP) {
+    public void setTopP(Double topP) {
         this.topP = topP;
     }
